InsertKey fix

+++ server/src/main/java/commands/noinput/Clear.java
@@ -5,6 +5,8 @@ import logic.Editor;
 import logic.InputData;
 import logic.OutputData;
 
+import java.sql.SQLException;
+
 public class Clear extends AbstractNoInputCommand {
     @Override
     public String getName() {
@@ -18,7 +20,15 @@ public class Clear extends AbstractNoInputCommand {
 
     @Override
     public OutputData exec(Editor editor, InputData inputData) {
-        editor.clear();
+        if (inputData.getAuth().equals("admin")) editor.clear();
+        else {
+            try {
+                editor.clear(inputData.getAuth());
+                return new OutputData("Success", "Successfully cleared your collection.");
+            } catch (SQLException exception) {
+                return new OutputData("Error", "Database problems....");
+            }
+        }
         return new OutputData("Success", "Successfully cleared the collection.");
     }
 }
